refactor: Moved FileOpenerExternal component
