+++ src/drive/web/modules/drive/index.jsx
@@ -1,4 +1,3 @@
-export { default as FileOpenerExternal } from './FileOpenerExternal.jsx'
 export { default as FolderContainer } from './FolderContainer.jsx'
 export { default as RecentContainer } from './RecentContainer.jsx'
 export { default as SharingsContainer } from './SharingsContainer.jsx'
