+++ src/drive/web/modules/navigation/AppRoute.jsx
@@ -8,11 +8,11 @@ import OnBoarding from 'drive/mobile/modules/onboarding/OnBoarding'
 import Layout from 'drive/web/modules/layout/Layout'
 import FileExplorer from './FileExplorer'
 import FilesViewer from 'drive/web/modules/viewer/FilesViewer'
+import FileOpenerExternal from 'drive/web/modules/viewer/FileOpenerExternal'
 
 import {
   FolderContainer as Folder,
   RecentContainer as Recent,
-  FileOpenerExternal,
   SharingsContainer as Sharings
 } from 'drive/web/modules/drive'
 import { Container as Trash } from 'drive/web/modules/trash'
